refactor: 레이스 및 결과 출력 하는 부분 외부 모듈로 분리
